requirements.txt
backcall==0.2.0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
cycler==0.10.0
decorator==5.0.9
idna==2.10
ipython==7.24.1
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.3.2
matplotlib-inline==0.1.2
nltk==3.6.2
numpy==1.20.3
pandas==1.2.4
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.18
ptyprocess==0.7.0
PyCap==1.1.3
Pygments==2.9.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.11.1
semantic-version==2.8.5
six==1.16.0
sklearn==0.0
threadpoolctl==2.1.0
tqdm==4.61.1
traitlets==5.0.5
urllib3==1.26.6
wcwidth==0.2.5
